% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_criteria.R
\name{conv.crit.loadings}
\alias{conv.crit.loadings}
\title{Calculate the maximum absolute difference in scaled loadings}
\usage{
conv.crit.loadings(new, old, k)
}
\arguments{
\item{new}{The \code{flash.fit} object from the current iteration.}

\item{old}{The \code{flash.fit} object from the previous iteration.}

\item{k}{Ignored.}
}
\description{
An alternative objective function that can be used to determine
  convergence when fitting a \code{\link{flash}} object. Calculates the
  maximum absolute difference in the L2-scaled loadings,
  \eqn{\max_{i, k} | \frac{\ell_ik^{new}}{\| \ell_k^{new} \|_2} -
  \frac{\ell_ik^{old}}{\| \ell_k^{old} \|_2} |}.
}
\seealso{
\code{\link{conv.crit.elbo}}, \code{\link{conv.crit.factors}}
}
