% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_info.R
\name{display.elbo}
\alias{display.elbo}
\title{Display the current ELBO}
\usage{
display.elbo(new, old, k)
}
\arguments{
\item{new}{The \code{flash.fit} object from the current iteration.}

\item{old}{The \code{flash.fit} object from the previous iteration.}

\item{k}{Only used during sequential backfits (that is, calls to
\code{\link{flash.backfit}} where \code{extrapolate = FALSE}). It then
gives the index of the factor/loadings pair currently being optimized.}
}
\description{
Displays the value of the variational lower bound (ELBO) at the current
  iteration.
}
\seealso{
\code{\link{display.elbo.diff}}, \code{\link{display.max.chg}},
  \code{\link{display.max.chg.L}}, \code{\link{display.max.chg.F}}
}
