% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_info.R
\name{display.max.chg}
\alias{display.max.chg}
\title{Display the maximum difference in scaled loadings and factors}
\usage{
display.max.chg(new, old, k)
}
\arguments{
\item{new}{The \code{flash.fit} object from the current iteration.}

\item{old}{The \code{flash.fit} object from the previous iteration.}

\item{k}{Only used during sequential backfits (that is, calls to
\code{\link{flash.backfit}} where \code{extrapolate = FALSE}). It then
gives the index of the factor/loadings pair currently being optimized.}
}
\description{
Displays the maximum (absolute) change over all L2-normalized loadings
  \eqn{\ell_1, \ldots, \ell_K} and factors \eqn{f_1, \ldots, f_K}.
}
\seealso{
\code{\link{display.elbo}}, \code{\link{display.elbo.diff}},
  \code{\link{display.L.max.chg}}, \code{\link{display.F.max.chg}}
}
