% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_fns.R
\name{ebnm.ash}
\alias{ebnm.ash}
\title{Wrapper to \code{ashr::ash}}
\usage{
ebnm.ash(x, s, g, fixg, output, ...)
}
\arguments{
\item{x}{A vector of observations.}

\item{s}{A vector of standard errors (or a scalar when all standard errors
are the same).}

\item{g}{An initialization for the prior.}

\item{fixg}{A boolean indicating whether \code{g} should be considered
fixed.}

\item{output}{Can be set to \code{default.output()} to return summary data
used to update factors; \code{"post_sampler"} to return a function that
will sample from the posterior; or \code{"lfsr"} to return a vector of
local false sign rates.}
}
\description{
\code{flashier} is designed to use \code{ebnm::ebnm} to solve the empirical
  Bayes normal means problem. This function wraps \code{ashr::ash} so that
  it can use the \code{ebnm} interface.
}
