% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_fns.R
\name{ebnm.pn}
\alias{ebnm.pn}
\alias{ebnm.ash}
\title{Flashier EBNM functions}
\usage{
ebnm.pn(x, s, g = NULL, fixg = FALSE, output = "flash.data", ...)

ebnm.ash(x, s, g = NULL, fixg = FALSE, output = "flash.data", ...)
}
\arguments{
\item{x}{A vector of observations.}

\item{s}{A vector of standard errors (or a scalar when all standard errors
are the same).}

\item{g}{An initialization for the prior.}

\item{fixg}{A boolean indicating whether \code{g} should be considered
fixed.}

\item{output}{Can be set to \code{"flash.data"} to return summary data used
to update factors; \code{"sampler"} to return a function that will sample
from the posterior; or \code{"lfsr"} to return a vector of local false
sign rates.}
}
\description{
Functions used to solve the empirical Bayes normal means problem.
  \code{ebnm.pn} is a wrapper to \code{ebnm::ebnm}. \code{ebnm.ash} calls
  \code{ashr::ash}.
}
\details{
Custom EBNM functions may also be used. They must accept parameters
  \code{x}, \code{s}, \code{g} (with default \code{g = NULL}),
  \code{fixg} (with default \code{fixg = FALSE}), and \code{output}. When
  \code{output = "flash.data"}, they should result a list containing
  elements \code{fitted_g} (the fitted prior), \code{loglik} (the
  log-likelihood for the observations given the prior), and \code{result},
  which must itself be a list containing elements \code{PosteriorMean}
  (the posterior means for the observations) and \code{PosteriorMean2}
  (the posterior expected squared values). When \code{output = "sampler"},
  they should return a list containing element \code{post_sampler} (a
  function that takes a single parameter \code{nsamp} and returns a matrix
  with \code{nsamp} rows and \code{length(x)} columns, with each row
  corresponding to a single sample from the posterior). Finally, when
  \code{output = "lfsr"}, they should return a list containing element
  \code{result}, which is itself a list containing element \code{lfsr},
  a vector of local false sign rates. For both \code{output = "sampler"}
  and \code{output = "lfsr"}, a \code{NULL} value may be returned without
  otherwise affecting the fitting process.
}
