% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_methods.R
\name{fitted.flash.fit}
\alias{fitted.flash.fit}
\title{Fitted method for flash fit objects}
\usage{
\method{fitted}{flash.fit}(object, ...)
}
\arguments{
\item{object}{An object inheriting from class \code{flash.fit}.}

\item{...}{Additional parameters are ignored.}
}
\description{
Give a \code{flash.fit} object, returns the "fitted values"
  \eqn{E(LF') = E(L) E(F)'}.
}
\seealso{
\code{\link{flash}}
}
