% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_methods.R
\name{fitted.flash}
\alias{fitted.flash}
\title{Fitted method for flash objects}
\usage{
\method{fitted}{flash}(object, ...)
}
\arguments{
\item{object}{An object inheriting from class \code{flash}.}

\item{...}{Additional parameters are ignored.}
}
\value{
The matrix of "fitted values."
}
\description{
Given a \code{\link{flash}} object, returns the "fitted values"
\eqn{E(LF') = E(L) E(F)'}.
}
