% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_methods.R
\name{fitted.flash_fit}
\alias{fitted.flash_fit}
\title{Fitted method for flash fit objects}
\usage{
\method{fitted}{flash_fit}(object, ...)
}
\arguments{
\item{object}{An object inheriting from class \code{flash_fit}.}

\item{...}{Additional parameters are ignored.}
}
\value{
The matrix of "fitted values."
}
\description{
Given a \code{\link{flash_fit}} object, returns the "fitted values"
\eqn{E(LF') = E(L) E(F)'}.
}
