% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_factor_types.R
\name{fixed.ones}
\alias{fixed.ones}
\alias{fixed.sparse}
\alias{fixed.factors}
\title{Flashier fixed factors}
\usage{
fixed.ones(n)

fixed.sparse(n, nz.idx)

fixed.factors(n, vals, idx = NULL)
}
\arguments{
\item{n}{The mode of the fixed loadings. For matrices, \code{n = 1}
indicates that row loadings will be fixed. \code{n = 2} designates
columns.}

\item{nz.idx}{For sparse factors, \code{nz.idx} gives the indices of the
nonzero loadings (that is, the loadings that are to be estimated rather
than fixed at zero). If \code{nz.idx} is a vector, then a single sparse
factor will be created. If it is a list of vectors, then multiple factors
will be created.}

\item{vals}{The values of the fixed loadings. Can be a vector, which is
interpreted as a single fixed factor, or a matrix, each column of which
is interpreted as a different fixed factor.}

\item{idx}{The indices of the fixed loadings. Set \code{idx = NULL} to fix
all mode-\code{n} loadings.}
}
\description{
Functions for creating different types of fixed factors. Note that loadings
  can only be fixed along one mode (rows, columns) at a time.
  \code{fixed.ones} creates a factor whose loadings along one mode are all
  fixed at 1. This type of factor can be used to estimate
  row and column means. \code{fixed.sparse} creates one or more factors
  with some loadings fixed at zero and with all other loadings to be
  estimated. \code{fixed.factors} is a more general function which creates
  one or more factors with loadings fixed at arbitrarily specified values.
}
