% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_fit_object.R
\name{flash.fit}
\alias{flash.fit}
\alias{ff.pm}
\alias{ff.p2m}
\alias{ff.var.type}
\alias{ff.fixed.var.type}
\alias{ff.est.tau}
\alias{ff.fixed.tau}
\alias{ff.tau}
\alias{ff.elbo}
\alias{ff.KL}
\alias{ff.g}
\alias{ff.data}
\alias{ff.nonmissing}
\alias{ff.R2}
\title{Extract a flash.fit object}
\usage{
flash.fit(flash)

ff.pm(f, n)

ff.p2m(f, n)

ff.var.type(f)

ff.fixed.var.type(f)

ff.est.tau(f)

ff.fixed.tau(f)

ff.tau(f)

ff.elbo(f)

ff.KL(f, k, n)

ff.g(f, k, n)

ff.data(f)

ff.nonmissing(f)

ff.R2(f)
}
\arguments{
\item{flash}{A \code{flash} object.}

\item{f}{A \code{flash.fit} object.}

\item{n}{The "mode" (set \code{n = 1} for loadings \eqn{L} and \code{n = 2}
for factors \eqn{F}).}

\item{k}{Only used during sequential backfits (that is, calls to
\code{flash.backfit} where \code{extrapolate = FALSE}). It then gives the
index of the factor/loadings pair currently being optimized.}
}
\description{
\code{flash.fit} objects are the "internal" objects used by \code{flash}
  functions to fit an EBMF model. Whereas \code{flash} objects
  (the end results of the fitting process) include user-friendly fields and
  methods, \code{flash.fit} objects were not designed for public
  consumption and can be unwieldy. Nonetheless, some advanced
  \code{flash} functionality requires the wielding of
  \code{flash.fit} objects. In particular, initialization, convergence,
  and "verbose" display functions all take one or more \code{flash.fit}
  objects as input (see parameter \code{init.fn} in function
  \code{\link{flash.add.greedy}}; parameter \code{conv.crit.fn} in
  \code{\link{flash.add.greedy}} and \code{\link{flash.backfit}};
  and parameter \code{disp.fns} in \code{\link{flash.set.verbose}}).
  For users who would like to write custom functions, the getter functions
  and methods enumerated below may prove useful.
}
\details{
The following S3 methods are available for \code{flash.fit} objects:
  \describe{
    \item{\code{\link{fitted.flash.fit}}}{Returns the "fitted values"
      \eqn{E(LF') = E(L) E(F)'}.}
    \item{\code{\link{residuals.flash.fit}}}{Returns the expected residuals
      \eqn{Y - E(LF') = Y - E(L) E(F)'}.}
    \item{\code{\link{ldf.flash.fit}}}{Returns an \eqn{LDF} decomposition,
      with columns of \eqn{L} and \eqn{F} scaled as specified by the user.}
  }
}
\section{Functions}{
\itemize{
\item \code{ff.pm}: The matrix of posterior means for the loadings matrix
\eqn{L} (when parameter \code{n} is equal to \code{1}) or factors matrix
\eqn{F} (when \code{n = 2}).

\item \code{ff.p2m}: The matrix of posterior second moments for the
loadings matrix \eqn{L} (when \code{n = 1}) or factors matrix \eqn{F}
(when \code{n = 2}).

\item \code{ff.var.type}: Describes the structure of the estimated portion of
the residual variance. Identical to parameter \code{var.type} in function
\code{\link{flash}}.

\item \code{ff.fixed.var.type}: Describes the structure of the fixed portion of
the residual variance. Analogous to \code{ff.var.type}.

\item \code{ff.est.tau}: Equal to \eqn{1 / \sigma^2}, where \eqn{\sigma^2}
is the estimated portion of the residual variance (total, by row, or
by column, depending on the variance type).

\item \code{ff.fixed.tau}: Equal to \eqn{1 / s^2}, where \eqn{s^2} is the
fixed portion of the residual variance (total, by row, or by column).

\item \code{ff.tau}: The overall precision \eqn{1 / (\sigma^2 + s^2)}.

\item \code{ff.elbo}: The variational lower bound (ELBO).

\item \code{ff.KL}: The KL-divergence portion of the ELBO for the
\eqn{k}th set of loadings (when \code{n = 1}) or factor (when
\code{n = 2}). During extrapolated backfits, a vector will be
returned with one value for each factor/loadings pair.

\item \code{ff.g}: The estimated prior \eqn{\hat{g}} for loadings
\eqn{\ell_k} (when \code{n = 1}) or for factor \eqn{f_k} (when
\code{n = 2}). During extrapolated backfits, a list of priors will be
returned with one element for each factor/loadings pair.

\item \code{ff.data}: The data matrix, with missing data (NAs) replaced by
zeros.

\item \code{ff.nonmissing}: A matrix of 0s and 1s indicating whether data
is missing (0) or present (1). If there is no missing data, then a scalar
might be stored rather than a full matrix of 1s.

\item \code{ff.R2}: The sum of expected squared residuals (total, by row,
or by column, depending on the variance type).
}}

