% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_fix_factors.R
\name{flash.fix.factors}
\alias{flash.fix.factors}
\title{Fix flash factors}
\usage{
flash.fix.factors(flash, kset, mode, is.fixed = TRUE, use.fixed.in.ebnm = NULL)
}
\arguments{
\item{flash}{A \code{flash} or \code{flash.fit} object}

\item{kset}{A vector of integers indexing the factor/loadings pairs whose
loadings or factors are to be fixed (or unfixed).}

\item{mode}{Set \code{mode = 1} to fix loadings and \code{mode = 2} to fix
factors.}

\item{is.fixed}{If \code{is.fixed = TRUE}, then all entries along the
specified mode will be fixed. If only a subset are to be fixed,
then \code{is.fixed} should be an appropriately-sized vector or
matrix of values that can be coerced to logical. For example, if
loadings for two factor/loadings pairs are to be fixed, then
\code{is.fixed} can be a length-\eqn{n} vector or an \eqn{n} by 2
matrix (where \eqn{n} is the number of rows in the data matrix).
Finally, loadings can be "unfixed" by setting \code{is.fixed = FALSE}.}

\item{use.fixed.in.ebnm}{By default, fixed elements are ignored when
solving the EBNM subproblem in order to estimate the prior \eqn{\hat{g}}.
This behavior can be changed by setting \code{use.fixed.in.ebnm = TRUE}.
This is a global setting which applies to all factor/loadings pairs;
behavior cannot vary from one pair to another.}
}
\value{
A \code{\link{flash}} object.
}
\description{
Fixes or unfixes loadings \eqn{\ell_k} or factors \eqn{f_k} for one or more
  factor/loadings pairs. For a given pair, either the loadings or factor can
  be fixed --- but not both ---, and either all loadings or a subset can be
  fixed. To unfix, set \code{is.fixed = FALSE}. See
  \code{\link{flash.init.factors}} for an example of usage.
}
