% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_init_factors.R
\name{flash.init.factors}
\alias{flash.init.factors}
\title{Initialize flash factors at specified values}
\usage{
flash.init.factors(flash, EF, EF2 = NULL,
  prior.family = prior.point.normal())
}
\arguments{
\item{flash}{A \code{flash} or \code{flash.fit} object to which factors are
to be added.}

\item{EF}{An SVD-like object or a list of matrices, one for each mode. Each
matrix should have k columns, one for each factor. For example, if the
data is a matrix of size n x p, then factors can be initialized by
supplying a list of length two, with the first element a matrix of size
n x k and the second element a matrix of size p x k. Missing entries are
not allowed.}

\item{EF2}{If NULL, then EF2 will be initialized at the element-wise squared
values of \code{EF}. Otherwise, a list of matrices (as described above)
must be supplied.}

\item{prior.family}{Indicates the family of distributions that the priors on
  the loadings are
  assumed to belong to. Can be a list of length 1 or length \eqn{N}, where
  \eqn{N} is the number of modes (\eqn{N = 2} for matrices; \eqn{N = 3} for
  tensors). Each list element must be a prior family defined by one of the
  convenience functions \code{\link{prior.normal}},
  \code{\link{prior.point.normal}},
  \code{\link{prior.point.laplace}}, \code{\link{prior.nonzero.mode}},
  \code{\link{prior.scale.normal.mix}}, \code{\link{prior.unimodal}},
  \code{\link{prior.symmetric.unimodal}},
  \code{\link{prior.nonnegative}}, or \code{\link{prior.nonpositive}},
  or a custom prior type of a similar form (see \code{\link{prior.normal}}
  for details).
  For example, the default \code{prior.family = prior.point.normal()} fits a
  (different) point-normal prior for each factor and each mode, while
  \code{prior.family = c(prior.nonnegative(), prior.scale.normal.mix())}
  fits a unimodal distribution with mode zero and nonnegative support to
  each set of row loadings and a scale mixture of normals with mean zero to
  each set of column loadings.

  \code{prior.family} can also be a list of lists, in which case the first
  list specifies the family or families for the first factor, the second
  specifies the family or families for the second factor, and so on. The
  last list element is then re-used as often as necessary.
  For example, \code{prior.family = list(prior.nonzero.mode(),
  prior.scale.normal.mix())} will fit point-normal priors with nonzero
  means for the first factor and scale mixtures of normals for every
  subsequent factor.}
}
\description{
Initializes factors at values specified by \code{EF} and \code{EF2}.
}
