% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_nullcheck.R
\name{flash.nullcheck}
\alias{flash.nullcheck}
\title{Nullcheck flash factors}
\usage{
flash.nullcheck(flash, kset = NULL, remove = FALSE,
  tol = set.default.tol(flash), verbose.lvl = get.verbose.lvl(flash))
}
\arguments{
\item{flash}{A \code{flash} or \code{flash.fit} object.}

\item{kset}{A vector of integers specifying which factors to nullcheck.
If \code{kset = NULL}, then all existing factors will be checked.}

\item{remove}{Whether to remove factors that have been set to zero from the
flash object. Note that this might change the indices of existing factors.}

\item{tol}{The tolerance parameter: if a factor does not improve the ELBO
by at least \code{tol}, then it will be set to zero.}

\item{verbose.lvl}{When and how to display progress updates. For nullchecks,
updates are only displayed when \code{verbose.lvl} > 0.}
}
\description{
Sets factors to zero if doing so improves the overall fit.
}
