% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_remove_factors.R
\name{flash.remove.factors}
\alias{flash.remove.factors}
\title{Remove factors from a flash object}
\usage{
flash.remove.factors(flash, kset)
}
\arguments{
\item{flash}{A \code{flash} or \code{flash.fit} object.}

\item{kset}{A vector of integers specifying which factors to remove.}
}
\description{
Sets flash factors to zero and then removes them from the flash object.
  Note that this will change the indices of existing flash factors.
}
