% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_reorder_factors.R
\name{flash.reorder.factors}
\alias{flash.reorder.factors}
\title{Reorder factors in a flash object}
\usage{
flash.reorder.factors(flash, kset)
}
\arguments{
\item{flash}{A \code{flash} or \code{flash.fit} object.}

\item{kset}{A vector of integers specifying the new order of the
factor/loadings pairs. All existing factors must be included in
\code{kset}; to drop factors, use \code{\link{flash.remove.factors}}.}
}
\value{
A \code{\link{flash}} object.
}
\description{
Reorders the factor/loadings pairs in a \code{flash} object.
}
