% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_remove_factors.R
\name{flash.set.factors.to.zero}
\alias{flash.set.factors.to.zero}
\title{Set flash factors to zero}
\usage{
flash.set.factors.to.zero(flash, kset)
}
\arguments{
\item{flash}{A \code{flash} or \code{flash.fit} object.}

\item{kset}{A vector of integers specifying which factors to remove.}
}
\description{
Sets flash factors to zero but does not remove them (so as to keep the
  indices of existing factors the same).
}
\seealso{
flash.remove.factors
}
