% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_set_verbose.R
\name{flash.set.verbose}
\alias{flash.set.verbose}
\title{Set verbose output}
\usage{
flash.set.verbose(flash, verbose, fns = NULL, colnames = NULL,
  colwidths = NULL)
}
\arguments{
\item{flash}{A \code{flash} or \code{flash.fit} object.}

\item{verbose}{When and how to display progress updates. Set to
\code{0} for none, \code{1} for updates after a factor is added or a
backfit is completed, \code{2} for additional notifications about the
variational lower bound, and \code{3} for updates after every iteration.
Set to \code{-1} to output a single tab-delimited table of values.}

\item{fns}{A vector of functions. Used to calculate values to output after
each factor update when \code{verbose} is either -1 or 3.}

\item{colnames}{A vector of column names, one for each function in
\code{verbose.fns}.}

\item{colwidths}{A vector of column widths.}
}
\description{
Sets the default verbosity level and verbose output for a flash object.
}
