% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_set_timeout.R
\name{flash_clear_timeout}
\alias{flash_clear_timeout}
\title{Set timeout}
\usage{
flash_clear_timeout(flash)
}
\arguments{
\item{flash}{A \code{flash} or \code{flash_fit} object.}
}
\value{
The \code{\link{flash}} object from argument \code{flash}, with
timeout settings cleared.
}
\description{
Used in a \code{\link{flash}} pipeline to clear timeout conditions set using
\code{\link{flash_set_timeout}}.
}
