% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_factors_remove.R
\name{flash_factors_remove}
\alias{flash_factors_remove}
\title{Remove factors from a flash object}
\usage{
flash_factors_remove(flash, kset)
}
\arguments{
\item{flash}{A \code{flash} or \code{flash_fit} object.}

\item{kset}{A vector of integers specifying which factor/loadings pairs to
remove.}
}
\value{
The \code{\link{flash}} object from argument \code{flash}, with the
factors specified by \code{kset} removed.
}
\description{
Sets factor/loadings pairs to zero and then removes them from the
\code{\link{flash}} object. Note that this will change the indices of
existing pairs.
}
\seealso{
\code{\link{flash_factors_set_to_zero}}
}
