% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_factors_reorder.R
\name{flash_factors_reorder}
\alias{flash_factors_reorder}
\title{Reorder factors in a flash object}
\usage{
flash_factors_reorder(flash, kset)
}
\arguments{
\item{flash}{A \code{flash} or \code{flash_fit} object.}

\item{kset}{A vector of integers specifying the new order of the
factor/loadings pairs. All existing factors must be included in
\code{kset}; to drop factors, use \code{\link{flash_factors_remove}}.}
}
\value{
The \code{\link{flash}} object from argument \code{flash}, with the
factors reordered according to argument \code{kset}.
}
\description{
Reorders the factor/loadings pairs in a \code{\link{flash}} object.
}
