% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_factors_remove.R
\name{flash_factors_set_to_zero}
\alias{flash_factors_set_to_zero}
\title{Set flash factors to zero}
\usage{
flash_factors_set_to_zero(flash, kset)
}
\arguments{
\item{flash}{A \code{flash} or \code{flash_fit} object.}

\item{kset}{A vector of integers specifying which factor/loadings pairs to
set to zero.}
}
\value{
The \code{\link{flash}} object from argument \code{flash}, with the
factors specified by \code{kset} set to zero.
}
\description{
Sets factor/loadings pairs to zero but does not remove them from the
\code{\link{flash}} object (so as to keep the indices of existing pairs
the same).
}
\seealso{
\code{\link{flash_factors_remove}}
}
