% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_factors_fix.R
\name{flash_factors_unfix}
\alias{flash_factors_unfix}
\title{Unfix flash factors}
\usage{
flash_factors_unfix(flash, kset)
}
\arguments{
\item{flash}{A \code{flash} or \code{flash_fit} object.}

\item{kset}{A vector of integers indexing the factor/loadings pairs whose
values are to be unfixed.}
}
\value{
The \code{\link{flash}} object from argument \code{flash}, with
values for the factor/loadings pairs specified by \code{kset} unfixed.
}
\description{
If loadings \eqn{\ell_{\cdot k}} or factors \eqn{f_{\cdot k}} for one or
more factor/loadings pairs have been "fixed" using function
\code{\link{flash_factors_fix}}, then they can be unfixed using
function \code{flash_factors_unfix}.
}
