% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_plots.R
\name{flash_plot_bar}
\alias{flash_plot_bar}
\title{Create bar plots of factors or loadings for a flash fit}
\usage{
flash_plot_bar(
  fl,
  order_by_pve = FALSE,
  kset = NULL,
  pm_which = c("factors", "loadings"),
  pm_subset = NULL,
  pm_groups = NULL,
  pm_colors = NULL,
  labels = FALSE,
  ...
)
}
\arguments{
\item{fl}{An object inheriting from class \code{flash}.}

\item{order_by_pve}{If \code{order_by_pve = TRUE}, then factor/loadings pairs
will be ordered according to proportion of variance explained, from
highest to lowest. (By default, they are plotted in the same order as
\code{kset}; or, if \code{kset} is \code{NULL}, then they are plotted in
the same order as they are found in \code{fl}.)}

\item{kset}{A vector of integers specifying the factor/loadings pairs to be
plotted. If \code{order_by_pve = FALSE}, then \code{kset} also specifies the
\emph{order} in which they are to be plotted.}

\item{pm_which}{Whether to plot loadings \eqn{L} or factors \eqn{F}.}

\item{pm_subset}{A vector of row indices \eqn{i} or column indices
\eqn{j} (depending on the argument to \code{pm_which})
specifying which values \eqn{\ell_{i \cdot}}{l_{i.}} or \eqn{f_{j \cdot}}{f_{j.}} are
to be shown. If the dataset has row or column names, then names rather
than indices may be specified. If \code{pm_subset = NULL}, then all values
will be plotted.}

\item{pm_groups}{A vector specifying the group to which each row of the data
\eqn{y_{i \cdot}}{y_{i.}} or column \eqn{y_{\cdot j}}{y_{.j}} belongs (groups may be
numeric indices or strings). A group must be provided for each plotted row
\eqn{i} or column \eqn{j}, so that the length of \code{pm_groups} is
exactly equal to the number of rows or columns in the full dataset or, if
\code{pm_subset} is specified, in the subsetted dataset.}

\item{pm_colors}{A character vector specifying a color for each unique group
specified by \code{pm_groups}, or, if \code{pm_groups = NULL}, a vector
specifying a color for each plotted row \eqn{i} or column \eqn{j}. Defines
the color (fill) of the bars.}

\item{labels}{Whether to label the bars along the \eqn{x}-axis. The
appearance of the labels (size, angle, etc.) can be adjusted using
\code{ggplot2}'s theme system; see below for an example.}

\item{...}{Additional arguments to be passed to
\code{\link[ggplot2]{facet_wrap}} (e.g., \code{nrow} or \code{ncol}).}
}
\value{
A \code{ggplot} object.
}
\description{
Creates a bar plot or sequence of bar plots, one for each value of \eqn{k} in
\code{kset}, with bars corresponding to individual posterior means for factors
\eqn{f_{jk}} or loadings \eqn{\ell_{ik}}. Values are normalized so that the
maximum absolute value for each factor \eqn{f_{\cdot k}} or set of
loadings \eqn{\ell_{\cdot k}} is equal to 1 (see \code{\link{ldf.flash}}).
This type of plot is most useful when rows \eqn{i = 1, \ldots, n} or columns
\eqn{j = 1, \ldots, p} are small in number or ordered in a logical fashion
(e.g., spatially).
}
\details{
When there is more than one value of \eqn{k} in \code{kset}, a sequence of
panels is created using the \code{\link[ggplot2]{facet_wrap}} function from
the \code{ggplot2} package. In each panel, the order of bars is determined
by the order of the corresponding rows or columns in the data matrix;
they can be re-arranged using the \code{pm_subset} argument.
}
\examples{
data(gtex)
fl <- flash(gtex, greedy_Kmax = 4L, backfit = FALSE)
flash_plot_bar(fl, pm_colors = gtex_colors)

# Tweaks are often required to get x-axis labels to look good:
library(ggplot2)
flash_plot_bar(fl, pm_colors = gtex_colors, labels = TRUE, ncol = 1) +
  theme(axis.text.x = element_text(size = 8, angle = 60))

}
