% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_plots.R
\name{flash_plot_heatmap}
\alias{flash_plot_heatmap}
\title{Create heatmap of factors or loadings for a flash fit}
\usage{
flash_plot_heatmap(
  fl,
  order_by_pve = FALSE,
  kset = NULL,
  pm_which = c("factors", "loadings"),
  pm_subset = NULL,
  pm_groups = NULL,
  pm_colors = NULL,
  gap = 1,
  ...
)
}
\arguments{
\item{fl}{An object inheriting from class \code{flash}.}

\item{order_by_pve}{If \code{order_by_pve = TRUE}, then factor/loadings pairs
will be ordered according to proportion of variance explained, from
highest to lowest. (By default, they are plotted in the same order as
\code{kset}; or, if \code{kset} is \code{NULL}, then they are plotted in
the same order as they are found in \code{fl}.)}

\item{kset}{A vector of integers specifying the factor/loadings pairs to be
plotted. If \code{order_by_pve = FALSE}, then \code{kset} also specifies the
\emph{order} in which they are to be plotted.}

\item{pm_which}{Whether to plot loadings \eqn{L} or factors \eqn{F}.}

\item{pm_subset}{A vector of row indices \eqn{i} or column indices
\eqn{j} (depending on the argument to \code{pm_which})
specifying which values \eqn{\ell_{i \cdot}}{l_{i.}} or \eqn{f_{j \cdot}}{f_{j.}} are
to be shown. If the dataset has row or column names, then names rather
than indices may be specified. If \code{pm_subset = NULL}, then all values
will be plotted.}

\item{pm_groups}{A vector specifying the group to which each row of the data
\eqn{y_{i \cdot}}{y_{i.}} or column \eqn{y_{\cdot j}}{y_{.j}} belongs (groups may be
numeric indices or strings). A group must be provided for each plotted row
\eqn{i} or column \eqn{j}, so that the length of \code{pm_groups} is
exactly equal to the number of rows or columns in the full dataset or, if
\code{pm_subset} is specified, in the subsetted dataset.}

\item{pm_colors}{A character vector of length 1, 2, or 3 defining the
diverging color gradient (low-mid-high) to be used by the heatmap. The
midpoint is set at zero. If one or two colors are supplied, then the
"mid" color will be set to white. If one color is supplied, then the "low"
and "high" colors (used for, respectively, negative and positive posterior
means) will be the same. If two are supplied, then the "low" color should
be provided first, followed by the "high" color. If all three are supplied,
then the "low" color should be provided first, followed by the "mid" color,
followed by the "high" color provided. The default color gradient is
\code{darkblue} for "low" (negative posterior means), white for "mid"
(zero), and \code{darkred} for "high" (positive posterior means).}

\item{gap}{The horizontal spacing between groups. Ignored if \code{pm_groups}
is not provided.}

\item{...}{Additional parameters to be passed to
\code{\link[fastTopics]{structure_plot}} (which is primarily used to
arrange the rows \eqn{i} or columns \eqn{j}).}
}
\value{
A \code{ggplot} object.
}
\description{
Creates a heatmap of posterior means for factors \eqn{f_{jk}} or loadings
\eqn{\ell_{ik}}{l_{ik}}. Values are normalized so that the maximum absolute value
for each factor \eqn{f_{\cdot k}}{f_{.k}} or set of loadings \eqn{\ell_{\cdot k}}{l_{.k}}
is equal to 1 (see \code{\link{ldf.flash}}).
}
\details{
By default, a 1-d embedding is used to arrange the rows \eqn{i} or columns
\eqn{j} in a "smart" manner. This behavior can be overridden via argument
\code{loadings_order}, which is passed to function
\code{\link[fastTopics]{structure_plot}}.
}
