% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_plots.R
\name{flash_plot_histogram}
\alias{flash_plot_histogram}
\title{Create histograms of factors or loadings for a flash fit}
\usage{
flash_plot_histogram(
  fl,
  order_by_pve = FALSE,
  kset = NULL,
  pm_which = c("factors", "loadings"),
  pm_subset = NULL,
  pm_groups = NULL,
  pm_colors = NULL,
  binwidth = NULL,
  bins = NULL,
  alpha = 0.5,
  ...
)
}
\arguments{
\item{fl}{An object inheriting from class \code{flash}.}

\item{order_by_pve}{If \code{order_by_pve = TRUE}, then factor/loadings pairs
will be ordered according to proportion of variance explained, from
highest to lowest. (By default, they are plotted in the same order as
\code{kset}; or, if \code{kset} is \code{NULL}, then they are plotted in
the same order as they are found in \code{fl}.)}

\item{kset}{A vector of integers specifying the factor/loadings pairs to be
plotted. If \code{order_by_pve = FALSE}, then \code{kset} also specifies the
\emph{order} in which they are to be plotted.}

\item{pm_which}{Whether to plot loadings \eqn{L} or factors \eqn{F}.}

\item{pm_subset}{A vector of row indices \eqn{i} or column indices
\eqn{j} (depending on the argument to \code{pm_which})
specifying which values \eqn{\ell_{i \cdot}}{l_{i.}} or \eqn{f_{j \cdot}}{f_{j.}} are
to be shown. If the dataset has row or column names, then names rather
than indices may be specified. If \code{pm_subset = NULL}, then all values
will be plotted.}

\item{pm_groups}{A vector specifying the group to which each row of the data
\eqn{y_{i \cdot}}{y_{i.}} or column \eqn{y_{\cdot j}}{y_{.j}} belongs (groups may be
numeric indices or strings). A group must be provided for each plotted row
\eqn{i} or column \eqn{j}, so that the length of \code{pm_groups} is
exactly equal to the number of rows or columns in the full dataset or, if
\code{pm_subset} is specified, in the subsetted dataset.}

\item{pm_colors}{A character vector specifying a color for each unique group
specified by \code{pm_groups}. Defines the color and fill of the histograms.}

\item{binwidth}{The width of the bins (a numeric value). The default is to
use the number of bins in \code{bins}, covering the range of the data.}

\item{bins}{Number of bins. Overriden by \code{binwidth}. Defaults to 30.}

\item{alpha}{A transparency value between 0 (transparent) and 1 (opaque).}

\item{...}{Additional arguments to be passed to
\code{\link[ggplot2]{facet_wrap}} (e.g., \code{nrow} or \code{ncol}).}
}
\value{
A \code{ggplot} object.
}
\description{
Creates a histogram or sequence of histograms of posterior means for factors
\eqn{f_{jk}} or loadings \eqn{\ell_{ik}}. One plot is created for each
value of \eqn{k} in \code{kset}. Values are normalized so that the
maximum absolute value for each factor \eqn{f_{\cdot k}} or set of
loadings \eqn{\ell_{\cdot k}} is equal to 1 (see \code{\link{ldf.flash}}).
If \code{pm_groups} is specified, then overlapping semi-transparent
histograms are created, with one histogram per group specified by
\code{pm_groups}. This option works best when the number of groups is small
or when groups are well separated across components.
}
