% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_plots.R
\name{flash_plot_scatter}
\alias{flash_plot_scatter}
\title{Create scatter plots of factors or loadings for a flash fit}
\usage{
flash_plot_scatter(
  fl,
  order_by_pve = FALSE,
  kset = NULL,
  pm_which = c("factors", "loadings"),
  pm_subset = NULL,
  pm_groups = NULL,
  pm_colors = NULL,
  covariate = NULL,
  shape = 1,
  labels = FALSE,
  n_labels = 0,
  label_size = 3,
  max_overlaps = Inf,
  ...
)
}
\arguments{
\item{fl}{An object inheriting from class \code{flash}.}

\item{order_by_pve}{If \code{order_by_pve = TRUE}, then factor/loadings pairs
will be ordered according to proportion of variance explained, from
highest to lowest. (By default, they are plotted in the same order as
\code{kset}; or, if \code{kset} is \code{NULL}, then they are plotted in
the same order as they are found in \code{fl}.)}

\item{kset}{A vector of integers specifying the factor/loadings pairs to be
plotted. If \code{order_by_pve = FALSE}, then \code{kset} also specifies the
\emph{order} in which they are to be plotted.}

\item{pm_which}{Whether to plot loadings \eqn{L} or factors \eqn{F}.}

\item{pm_subset}{A vector of row indices \eqn{i} or column indices
\eqn{j} (depending on the argument to \code{pm_which})
specifying which values \eqn{\ell_{i \cdot}}{l_{i.}} or \eqn{f_{j \cdot}}{f_{j.}} are
to be shown. If the dataset has row or column names, then names rather
than indices may be specified. If \code{pm_subset = NULL}, then all values
will be plotted.}

\item{pm_groups}{A vector specifying the group to which each row of the data
\eqn{y_{i \cdot}}{y_{i.}} or column \eqn{y_{\cdot j}}{y_{.j}} belongs (groups may be
numeric indices or strings). A group must be provided for each plotted row
\eqn{i} or column \eqn{j}, so that the length of \code{pm_groups} is
exactly equal to the number of rows or columns in the full dataset or, if
\code{pm_subset} is specified, in the subsetted dataset.}

\item{pm_colors}{A character vector specifying a color for each unique group
specified by \code{pm_groups}, or, if \code{pm_groups = NULL}, a vector
specifying a color for each plotted row \eqn{i} or column \eqn{j}. Defines
the colors of the points.}

\item{covariate}{A numeric vector with one value for each plotted row \eqn{i}
or column \eqn{j}. These values are mapped onto the plots' \eqn{y}-axis.}

\item{shape}{The symbol used for the plots' points. See
\code{\link[ggplot2]{aes_linetype_size_shape}}.}

\item{labels}{Whether to label the points with the largest (absolute)
posterior means. If \code{labels = TRUE}, then \code{n_labels} points will
be labelled using \code{\link[ggrepel]{geom_text_repel}}.}

\item{n_labels}{A (nonnegative) integer. The number of points to label. If
\code{n_labels} is set to a positive integer but \code{labels = FALSE},
then the \code{n_labels} points with the largest (absolute) posterior
means will be highlighted in blue but not labelled. This can be useful for
tweaking labels using the full range of options provided by
\code{\link[ggrepel]{geom_text_repel}}. For an example, see below.}

\item{label_size}{The size of the label text (in millimeters).}

\item{max_overlaps}{A (nonnegative) integer. For each text label, the number
of overlaps with other text labels or other data points are counted, and
the text label is excluded if it has too many overlaps.}

\item{...}{Additional arguments to be passed to
\code{\link[ggplot2]{facet_wrap}} (e.g., \code{nrow} or \code{ncol}).}
}
\value{
A \code{ggplot} object.
}
\description{
Creates a scatter plot or sequence of scatter plots, with position along the
\eqn{x}-axis defined by posterior means for factors \eqn{f_{jk}} or loadings
\eqn{\ell_{ik}} and position along the \eqn{y}-axis defined by a
user-supplied covariate. If a covariate is not supplied, then plots will
use data column or row means, \eqn{\frac{1}{n} \sum_{i = 1}^n y_{ij}} or
\eqn{\frac{1}{p} \sum_{j = 1}^p y_{ij}}. One plot is created for
each value of \eqn{k} in \code{kset}. Values are normalized so that the
maximum absolute value for each factor \eqn{f_{\cdot k}} or set of
loadings \eqn{\ell_{\cdot k}} is equal to 1 (see \code{\link{ldf.flash}}).
}
\examples{
data(gtex)
fl <- flash(gtex, greedy_Kmax = 4L, backfit = FALSE)
flash_plot_scatter(fl)

# Label axes and points:
library(ggplot2)
flash_plot_scatter(fl, labels = TRUE, n_labels = 3) +
  labs(y = "mean z-score across all SNPs")

# For the full range of labelling options provided by the ggrepel package, set
#   labels = FALSE (the default setting) and add geom_text_repel() manually:
library(ggrepel)
flash_plot_scatter(fl, labels = FALSE, n_labels = 3) +
  geom_text_repel(size = 2.5, min.segment.length = 0)

}
