% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_update_data.R
\name{flash_update_data}
\alias{flash_update_data}
\title{Update data in a flash object}
\usage{
flash_update_data(flash, newdata)
}
\arguments{
\item{flash}{A \code{flash} or \code{flash_fit} object.}

\item{newdata}{The new observations. Can be a matrix, a sparse matrix of
class \code{\link[Matrix]{Matrix}}, or a low-rank matrix representation.}
}
\value{
The \code{\link{flash}} object from argument \code{flash}, with
  the data modified as specified by \code{newdata}. Residual variances and
  ELBO are also updated.
}
\description{
Replaces the data in a flash object with a new set of observations. Estimates
  of residual variances and the ELBO are also updated.
}
