% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_fns.R
\name{init.fn.default}
\alias{init.fn.default}
\title{Initialize a flash factor}
\usage{
init.fn.default(flash, tol = 1/max(get.dims(flash)), maxiter = 100, seed = 666)
}
\arguments{
\item{flash}{A \code{flash.fit} object.}

\item{tol}{Convergence tolerance.}

\item{maxiter}{Maximum number of iterations.}

\item{seed}{Since initialization is random, a default seed is set for
reproducibility.}
}
\description{
Initializes a new flash factor. Custom initialization functions can also be
  used. They should accept a single parameter \code{flash} and
  output a list consisting of two vectors (which will be interpreted as a
  rank-one matrix).
}
\examples{

# Change the default initialization maxiter
my.init.fn <- function(flash) {
  return(init.fn.default(flash, maxiter = 500))
}

fl <- flash.init(gtex) \%>\% flash.add.greedy(init.fn = my.init.fn)

# A wrapper to function nnmf in package NNLM
nnmf.init.fn <- function(flash) {
  R <- flash$Y - flashier:::lowrank.expand(flash$EF)
  res <- NNLM::nnmf(R, verbose = FALSE)
  return(list(as.vector(res$W), as.vector(res$H)))
}

fl.nnmf <- flash.init(gtex) \%>\%
  flash.add.greedy(ebnm.fn = ebnm::ebnm_unimodal_nonnegative,
                   init.fn = nnmf.init.fn)

}
\seealso{
init.fn.softImpute
}
