% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_fns.R
\name{init.fn.irlba}
\alias{init.fn.irlba}
\title{Initialize a flash factor using IRLBA}
\usage{
init.fn.irlba(flash, seed = 666, ...)
}
\arguments{
\item{flash}{A \code{flash.fit} object.}

\item{seed}{Since initialization is random, a default seed is set for
reproducibility.}

\item{...}{Additional parameters to be passed to \code{\link[irlba]{irlba}}.}
}
\description{
Initializes a new flash factor using \code{\link[irlba]{irlba}}. This
  can be somewhat faster than \code{\link{init.fn.default}} for large,
  dense data matrices. For sparse matrices of class \code{Matrix}, the
  default initialization should generally be preferred.
}
\seealso{
\code{\link{init.fn.default}}, \code{\link{init.fn.softImpute}}
}
