% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_fns.R
\name{init.fn.softImpute}
\alias{init.fn.softImpute}
\title{Initialize a flash factor using softImpute}
\usage{
init.fn.softImpute(flash, seed = 666, ...)
}
\arguments{
\item{flash}{A \code{flash.fit} object.}

\item{seed}{Since initialization is random, a default seed is set for
reproducibility.}

\item{...}{Additional parameters to be passed to
\code{\link[softImpute]{softImpute}}.}
}
\description{
Initializes a new flash factor using \code{\link[softImpute]{softImpute}}.
  When there is missing data, this can yield better results than
  \code{\link{init.fn.default}} without sacrificing much (if any) speed.
}
\seealso{
\code{\link{init.fn.default}}, \code{\link{init.fn.irlba}}
}
