% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_methods.R
\name{ldf}
\alias{ldf}
\alias{ldf.flash}
\alias{ldf.flash.fit}
\title{LDF method for flash and flash fit objects}
\usage{
ldf(object, ...)

\method{ldf}{flash}(object, type = "f")

\method{ldf}{flash.fit}(object, type = "f")
}
\arguments{
\item{object}{An object inheriting from class \code{flash} or
\code{flash.fit}.}

\item{type}{Takes identical arguments to function \code{\link[base]{norm}}. Use
\code{"f"} or \code{"2"} for the 2-norm (Euclidean norm); \code{"o"} or
\code{"1"} for the 1-norm (taxicab norm); and \code{"i"} or \code{"m"} for
the infinity norm (maximum norm).}
}
\description{
Given a \code{flash} or \code{flash.fit} object, returns the LDF
  decomposition \eqn{Y \approx LDF'}.
}
\details{
When the prior families \eqn{G_\ell^{(k)}} and \eqn{G_f^{(k)}} are closed
  under scaling (as is typically the case), then the EBMF model (as
  described in the documention to function \code{\link{flash}}) is only
  identifiable up to scaling by a diagonal matrix \eqn{D}:
  \deqn{Y = LDF' + E.}

Method \code{ldf} scales columns \eqn{\ell_k} and \eqn{f_k}
  so that, depending on the argument to parameter \code{type}, their
  1-norms, 2-norms, or infinity norms are equal to 1.
}
\section{Methods (by class)}{
\itemize{
\item \code{ldf(flash)}: LDF decomposition for flash objects

\item \code{ldf(flash.fit)}: LDF decomposition for flash fit objects

}}
\seealso{
\code{\link{flash}}
}
