% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_methods.R
\name{plot.flash}
\alias{plot.flash}
\title{Plot method for flash objects}
\usage{
\method{plot}{flash}(
  x,
  incl.scree = TRUE,
  incl.pm = TRUE,
  order.by.pve = TRUE,
  kset = NULL,
  pm.which = c("factors", "loadings"),
  pm.subset = NULL,
  pm.groups = NULL,
  pm.colors = NULL,
  ...
)
}
\arguments{
\item{x}{An object inheriting from class \code{flash}.}

\item{incl.scree}{Whether to include a figure showing the proportion of
variance explained by each factor/loading pair ("scree plot").}

\item{incl.pm}{Whether to include a figure showing the posterior means for
either loadings \eqn{L} or factors \eqn{F} (depending on the argument to
\code{pm.which}). If argument \code{pm.groups}
is left unspecified, then bar plots will be produced, with each bar
corresponding to a single value \eqn{\ell_{ik}} or \eqn{f_{jk}}.
Otherwise, overlapping histograms will be
produced, with each histogram corresponding to one of the groups
specified by \code{pm.groups}. One plot is produced for each
factor/loading pair \eqn{k}.}

\item{order.by.pve}{If \code{TRUE}, then the factor/loading pairs will be
re-ordered according to proportion of variance explained (from
highest to lowest).}

\item{kset}{A vector of integers specifying the factor/loading pairs to be
plotted. If \code{kset = NULL}, then all will be plotted.}

\item{pm.which}{Whether to plot loadings \eqn{L} or factors \eqn{F} in the
plots of posterior means.}

\item{pm.subset}{A vector of loading indices \eqn{i} or factor indices
\eqn{j} (depending on the argument to \code{pm.which})
specifying which values \eqn{\ell_{i \cdot}} or \eqn{f_{j \cdot}} are
to be shown in the plots of posterior means. If the dataset has row or
column names, then names rather than indices may be specified. If
\code{pm.subset = NULL}, then all values will be plotted.}

\item{pm.groups}{A vector specifying the group to which each row
\eqn{y_{i \cdot}} or column \eqn{y_{\cdot j}} of the data belongs
(groups may be numeric indices or strings). If \code{pm.groups = NULL},
then a bar plot of the ungrouped data will be produced (see \code{incl.pm}
above). Otherwise, a group must be provided for each observation, so that
the length of \code{pm.groups} is exactly equal to the number of rows or
columns in the full dataset or, if \code{pm.subset} is specified, in the
subsetted dataset. If \code{pm.groups} is not \code{NULL}, then a set of
overlapping histograms will be produced for each factor/loading pair, with
one histogram per group (again see \code{incl.pm}).}

\item{pm.colors}{A vector specifying a color for each bar (if
\code{pm.groups = NULL}) or histogram (if \code{pm.groups} is not
\code{NULL}). Passed directly to argument \code{values} in \strong{ggplot2}
function \code{\link[ggplot2]{scale_color_manual}}.}

\item{...}{Additional parameters are ignored.}
}
\description{
Given a \code{\link{flash}} object, produces up to two figures: one showing
  the proportion of variance explained per factor, and one that plots
  posterior means.
}
\seealso{
\code{\link{flash}}
}
