% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_methods.R
\name{residuals.flash.fit}
\alias{residuals.flash.fit}
\title{Residuals method for flash fit objects}
\usage{
\method{residuals}{flash.fit}(object, ...)
}
\arguments{
\item{object}{An object inheriting from class \code{flash.fit}.}

\item{...}{Additional parameters are ignored.}
}
\description{
Given a \code{flash.fit} object, returns the expected residuals
  \eqn{Y - E(LF') = Y - E(L) E(F)'}.
}
\seealso{
\code{\link{flash}}
}
