% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_methods.R
\name{residuals.flash}
\alias{residuals.flash}
\title{Residuals method for flash objects}
\usage{
\method{residuals}{flash}(object, ...)
}
\arguments{
\item{object}{An object inheriting from class \code{flash}.}

\item{...}{Additional parameters are ignored.}
}
\value{
The matrix of expected residuals.
}
\description{
Given a \code{\link{flash}} object, returns the expected residuals
\eqn{Y - E(LF') = Y - E(L) E(F)'}.
}
