% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_methods.R
\name{residuals.flash_fit}
\alias{residuals.flash_fit}
\title{Residuals method for flash fit objects}
\usage{
\method{residuals}{flash_fit}(object, ...)
}
\arguments{
\item{object}{An object inheriting from class \code{flash_fit}.}

\item{...}{Additional parameters are ignored.}
}
\value{
The matrix of expected residuals.
}
\description{
Given a \code{\link{flash_fit}} object, returns the expected residuals
\eqn{Y - E(LF') = Y - E(L) E(F)'}.
}
