% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_fit_object.R
\name{flash.fit}
\alias{flash.fit}
\alias{ff.pm}
\alias{ff.p2m}
\alias{ff.est.tau}
\alias{ff.fixed.tau}
\alias{ff.tau}
\alias{ff.elbo}
\alias{ff.KL}
\alias{ff.g}
\title{Extract a flash.fit object}
\usage{
flash.fit(flash)

ff.pm(f, n)

ff.p2m(f, n)

ff.est.tau(f)

ff.fixed.tau(f)

ff.tau(f)

ff.elbo(f)

ff.KL(f, n)

ff.g(f, n)
}
\arguments{
\item{flash}{A \code{flash} object.}

\item{f}{A \code{flash.fit} object.}

\item{n}{The dimension (set \code{n = 1} for loadings \eqn{L} and \code{n = 2}
for factors \eqn{F}).}
}
\description{
\code{flash.fit} objects are the "internal" objects used by \code{flash}
  functions to fit an EBMF model. Whereas \code{flash} objects
  (the end results of the fitting process) include user-friendly fields and
  methods, \code{flash.fit} objects were not designed for public
  consumption and can be unwieldy. Nonetheless, some advanced
  \code{flash} functionality requires the wielding of
  \code{flash.fit} objects. In particular, initialization, convergence,
  and "verbose" display functions all take one or more \code{flash.fit}
  objects as input (see parameter \code{init.fn} in function
  \code{\link{flash.add.greedy}}; parameter \code{conv.crit.fn} in
  \code{\link{flash.add.greedy}} and \code{\link{flash.backfit}};
  and parameter \code{disp.fns} in \code{\link{flash.set.verbose}}).
  For users who would like to write custom functions, the getter functions
  and methods enumerated below may prove useful.
}
\details{
The following S3 methods are available for \code{flash.fit} objects at all
  times except while optimizing new factor/loadings pairs as part of a
  "greedy" fit:
  \describe{
    \item{\code{\link{fitted.flash.fit}}}{Returns the "fitted values"
      \eqn{E(LF') = E(L) E(F)'}.}
    \item{\code{\link{residuals.flash.fit}}}{Returns the expected residuals
      \eqn{Y - E(LF') = Y - E(L) E(F)'}.}
    \item{\code{\link{ldf.flash.fit}}}{Returns an \eqn{LDF} decomposition,
      with columns of \eqn{L} and \eqn{F} scaled as specified by the user.}
  }
}
\section{Functions}{
\itemize{
\item \code{ff.pm}: The posterior means for the loadings matrix \eqn{L}
(when parameter \code{n} is equal to \code{1}) or factor matrix \eqn{F}
(when \code{n = 2}). While optimizing new factor/loadings pairs as part of
a "greedy" fit, only the posterior means for the new loadings \eqn{\ell_k}
or factor \eqn{f_k} will be returned.

\item \code{ff.p2m}: The posterior second moments for the loadings matrix
\eqn{L} (when parameter \code{n} is equal to \code{1}) or factor matrix
\eqn{F} (when \code{n = 2}). While optimizing new factor/loadings pairs,
only the posterior second moments for the new loadings \eqn{\ell_k} or
factor \eqn{f_k} will be returned.

\item \code{ff.est.tau}: Equal to \eqn{1 / \sigma^2}, where \eqn{\sigma^2}
is the estimated portion of the residual variance (total, by row, or
by column, depending on the variance type).

\item \code{ff.fixed.tau}: Equal to \eqn{1 / s^2}, where \eqn{s^2} is the
fixed portion of the residual variance (total, by row, or by column).

\item \code{ff.tau}: The overall precision \eqn{1 / (\sigma^2 + s^2)}.

\item \code{ff.elbo}: The variational lower bound (ELBO).

\item \code{ff.KL}: A vector containing the KL-divergence portions of
the ELBO, with one value for each factor (when \code{n = 2}) or set of
loadings (when \code{n = 1}). While optimizing new factor/loadings pairs,
only the KL-divergence for the new loadings or factor will be returned.

\item \code{ff.g}: A list containing estimated priors on loadings
\eqn{\hat{g}_\ell} (when \code{n = 1}) or on factors \eqn{\hat{g}_f} (when
\code{n = 2}). While optimizing new factor/loadings pairs, only the
estimated prior on the new loadings or factor will be returned.
}}

