% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_remove_factors.R
\name{flash.set.factors.to.zero}
\alias{flash.set.factors.to.zero}
\title{Set flash factors to zero}
\usage{
flash.set.factors.to.zero(flash, kset)
}
\arguments{
\item{flash}{A \code{flash} or \code{flash.fit} object.}

\item{kset}{A vector of integers specifying which factors to remove.}
}
\value{
A \code{\link{flash}} object.
}
\description{
Sets factor/loading pairs to zero but does not remove them from the
  \code{flash} object (so as to keep the indices of existing pairs the same).
}
\seealso{
\code{\link{flash.remove.factors}}
}
