% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_set_verbose.R
\name{flash.set.verbose}
\alias{flash.set.verbose}
\title{Set verbose output}
\usage{
flash.set.verbose(flash, verbose, fns = NULL, colnames = NULL,
  colwidths = NULL)
}
\arguments{
\item{flash}{A \code{flash} or \code{flash.fit} object.}

\item{verbose}{When and how to display progress updates. Set to
  \code{0} for none, \code{1} for updates after a factor is added or a
  backfit is completed, \code{2} for additional notifications about the
  variational lower bound, and \code{3} for updates after every iteration.
  Set to \code{-1} to output a single tab-delimited table of values.

When \code{verbose = 3}, the improvement in the ELBO and the largest
  change in the loadings is shown after each iteration. If \code{verbose}
  is a character string rather than an integer, then a different set of
  columns will be shown. The character string should be a set of letters
  separated by spaces. Each letter may be followed by a number to specify
  a mode (e.g., rows or columns). Current options are: the (O)bjective
  after each iteration; the (D)ifference in the objective after each
  iteration; the maximum change in the mode-n (L)oadings after each
  iteration; the index corresponding to (W)hich mode-n loadings experienced
  the largest change; and the (S)parsity of the mode-n loadings after each
  iteration. See below for examples.

Note that output is completely customizable via parameters \code{fns},
  \code{colnames}, and \code{colwidths}. If these are used, then
  \code{verbose} must take an integer argument (and the specified columns
  will only display when \code{verbose} is either -1 or 3).}

\item{fns}{A vector of functions. Used to calculate values to output after
each factor update when \code{verbose} is either -1 or 3.}

\item{colnames}{A vector of column names, one for each function in
\code{fns}.}

\item{colwidths}{A vector of column widths.}
}
\description{
Sets the default verbosity level and verbose output for a flash object.
}
\examples{
data(gtex)

# Suppress all verbose output.
fl <- flash.init(gtex) \%>\%
  flash.set.verbose(0) \%>\%
  flash.add.greedy(Kmax = 5L)

# Set verbose output using a character string.
fl <- flash.init(gtex) \%>\%
  flash.set.verbose("D L2 W2 S1") \%>\%
  flash.add.greedy(Kmax = 5L)

# Set custom verbose output.
fns <- c(flashier:::calc.obj.diff,
         function(new, old, k) {
           flashier:::calc.max.abs.chg.EF(new, old, k, n = 2)
         })
colnames <- c("ELBO Diff", "Max Chg (Tiss)")
colwidths <- c(12, 18)
fl <- flash.init(gtex) \%>\%
  flash.set.verbose(3L, fns = fns, colnames = colnames, colwidths = colwidths) \%>\%
  flash.add.greedy(Kmax = 3L)

# Default output can be set and then turned off as needed.
fl <- flash.init(gtex) \%>\%
  flash.set.verbose("D L2 W2 S1") \%>\%
  flash.add.greedy(Kmax = 5L) \%>\%
  flash.backfit(verbose.lvl = 1) \%>\%
  flash.add.greedy(Kmax = 1L)

}
