% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_ebnm.R
\name{flash_ebnm}
\alias{flash_ebnm}
\title{Construct an EBNM function}
\usage{
flash_ebnm(...)
}
\arguments{
\item{...}{Parameters to be passed to function \code{\link[ebnm]{ebnm}}
in package \code{ebnm}. An argument to \code{prior_family} should be
provided unless the default family of point-normal priors is desired.
Arguments to parameters \code{x}, \code{s}, or \code{output} must not be
included. Finally, if \code{g_init} is included, then \code{fix_g = TRUE}
must be as well. To fix a prior grid, use parameter \code{scale} rather
than \code{g_init}.}
}
\value{
A function that can be passed as argument to parameter
\code{ebnm_fn} in functions \code{\link{flash}},
\code{\link{flash_greedy}}, and \code{\link{flash_factors_init}}.
}
\description{
\code{flash_ebnm} is a helper function that provides readable syntax for
constructing \code{\link[ebnm]{ebnm}} functions that can serve as
arguments to parameter \code{ebnm_fn} in functions \code{\link{flash}},
\code{\link{flash_greedy}}, and \code{\link{flash_factors_init}} (see
\strong{Examples} below). It is also possible to write a custom function
from scratch: see \strong{Details} below for a simple example. A more
involved example can be found in the "Extending ebnm with custom ebnm-style
functions" vignette in the \code{\link[ebnm]{ebnm}} package.
}
\details{
As input to parameter \code{ebnm_fn} in functions \code{\link{flash}},
\code{\link{flash_greedy}}, and \code{\link{flash_factors_init}},
it should suffice for many purposes to
provide functions from package \code{ebnm} as is (for example, one might
set \code{ebnm_fn = ebnm_point_laplace}). To use non-default
arguments, function \code{flash_ebnm} may be used (see \strong{Examples}).
Custom functions may also be written; for details, see the "Extending
ebnm with custom ebnm-style functions" vignette in the
\code{\link[ebnm]{ebnm}} package.
}
\examples{
# A custom EBNM function might be written as follows:
my_ebnm_fn <- function(x, s, g_init, fix_g, output) {
  ebnm_res <- ebnm_point_laplace(
    x = x,
    s = s,
    g_init = g_init,
    fix_g = fix_g,
    output = output,
    control = list(iterlim = 10)
  )
  return(ebnm_res)
}

# The following are equivalent:
fl1 <- flash(
  gtex,
  ebnm_fn = my_ebnm_fn,
  greedy_Kmax = 2
)
fl2 <- flash(
  gtex,
  ebnm_fn = flash_ebnm(
    prior_family = "point_laplace",
    control = list(iterlim = 10)
  ),
  greedy_Kmax = 2
)

}
\seealso{
\code{\link[ebnm]{ebnm}}
}
