% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_plots.R
\name{flash_plot_scree}
\alias{flash_plot_scree}
\title{Create a scree plot for a flash fit}
\usage{
flash_plot_scree(fl, order_by_pve = FALSE, kset = NULL, labels = FALSE)
}
\arguments{
\item{order_by_pve}{If \code{TRUE}, then the factor/loadings pairs will be
re-ordered according to proportion of variance explained (from
highest to lowest).}

\item{kset}{A vector of integers specifying the factor/loadings pairs to be
plotted. If \code{kset = NULL}, then all will be plotted.}

\item{labels}{Whether to label the points in the scree plot with the
indices of the factor/loading pairs they correspond to. Labels appear
as "k1", "k2", "k3", etc.}
}
\value{
A \code{ggplot2} object.
}
\description{
A scree plot is a line plot showing the proportion of variance explained by
each factor/loadings pair in a flash fit. Note that since EBMF does not
require factors and loadings to be orthogonal, "PVE" should be
interpreted loosely: for example, the total proportion of variance
explained could be larger than 1.
}
\details{
Unlike scree plots for PCA, a scree plot for a flash fit is not in general
monotonically decreasing. To ensure a monotonically decreasing scree
plot, set \code{order_by_pve = TRUE}. Note, however, that if this is done
then the numbers on the \eqn{x}-axis will no longer match the indices of
the components in the flash fit. This can also be true if argument
\code{kset} has been specified. Thus one should consider setting
\code{labels = TRUE} when \code{order_by_pve = TRUE} or when \code{kset}
has been specified.
}
