% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashier.R
\name{flashier}
\alias{flashier}
\title{Fit an empirical Bayes matrix/tensor factorization}
\usage{
flashier(data = NULL, S = NULL, flash.init = NULL, var.type = 0,
  prior.type = "point.normal", greedy.Kmax = 30, backfit = c("none",
  "final", "alternating", "only"), ebnm.param = list(), ash.param = list(),
  verbose.lvl = 1, ...)
}
\arguments{
\item{data}{The observations. Can be a matrix, sparse matrix of class
\code{Matrix}, three-dimensional array, or \code{flash.data} object
obtained from \code{set.flash.data}. Can be \code{NULL} if
\code{flash.init} is used.}

\item{S}{The standard errors. Can be a matrix, scalar (if standard errors
are the same for all observations), or vector (if, for example, \code{S}
only varies across columns and is constant within any given row). If
\code{NULL}, all residual variance will be estimated.}

\item{flash.init}{An initial \code{flash} or \code{flash.fit} object.}

\item{var.type}{Describes the structure of the estimated residual variance.
Can be \code{NULL}, \code{0}, or a vector. If \code{NULL}, then
\code{S} accounts for all residual variance. Otherwise, a rank-one
variance structure will be estimated (and added to any variance specified
by \code{S}). \code{var.type} then gives the dimensions
along which the residual variance is permitted to vary. For example,
\code{var.type = 1} estimates row-specific residual variances, while
\code{var.type = c(1, 2)} estimates an arbitrary rank-one matrix. If
\code{var.type = 0}, then the residual variance is assumed to be constant
across all observations.}

\item{prior.type}{Indicates the class of distributions that the priors are
assumed to belong to. Can be a vector of length 1, a vector with one
element for each dimension, or a list of vectors. In the last case,
the first list element gives the class(es) for the first factor, the
second gives the class(es) for the second factor, and so on. The last
list element is then re-used as often as necessary. Options include
\code{"point.normal"}, \code{"point.laplace"}, \code{"normal.mixture"},
\code{"uniform.mixture"}, \code{"nonnegative"}, \code{"nonpositive"}, and
\code{"nonzero.mode"}.}

\item{greedy.Kmax}{The maximum number of factors to be added. Fixed factors
are not counted.}

\item{backfit}{Whether and how to backfit. If \code{"final"}, then a single
backfit is performed after as many factors as possible have been added.
If \code{"alternating"}, a backfit will be performed after each
factor is added. Set \code{backfit = "only"} to backfit \code{flash.init}
without adding additional factors.}

\item{ebnm.param}{Additional parameters to be passed to \code{ebnm::ebnm}.
Used by prior types \code{"point.normal"}, \code{"point.laplace"}, and
\code{"nonzero.mode"}.}

\item{ash.param}{Additional parameters to be passed to \code{ashr::ash}.
Used by prior types \code{"normal.mixture"}, \code{"uniform.mixture"},
\code{"nonnegative"}, and \code{"nonpositive"}.}

\item{verbose.lvl}{When and how to display progress updates. Set to
\code{0} for none, \code{1} for updates after a factor is added or a
backfit is completed, \code{2} for additional notifications about the
variational lower bound, and \code{3} for updates after every iteration.
Set to \code{-1} to output a single tab-delimited table of values.}

\item{...}{Additional parameters to be passed to
\code{\link{flash.workhorse}}. Not for the faint of heart.}
}
\value{
A \code{flash} object. Contains elements:
  \describe{
    \item{\code{n.factors}}{The total number of factors in the fitted
      model.}
    \item{\code{objective}}{The variational lower bound achieved by the
      fitted model.}
    \item{\code{pve}}{The proportion of variance explained by each factor.}
    \item{\code{loadings}}{For each factor, the normalized loadings and a
      scaling constant.}
    \item{\code{sampler}}{A function that takes a single argument
      \code{nsamp} and returns \code{nsamp} samples from the posterior
      distribution of the (non-normalized) loadings.}
    \item{\code{fit}}{A \code{flash.fit} object.}
  }
}
\description{
TODO: Describe model. Add reference.
}
