% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_fns.R
\name{init.fn.softImpute}
\alias{init.fn.softImpute}
\title{Initialize a flash factor using softImpute}
\usage{
init.fn.softImpute(
  flash,
  mode.signs,
  tol = 1/max(get.dims(flash)),
  maxiter = 100,
  seed = 666
)
}
\arguments{
\item{flash}{A \code{flash.fit} object.}

\item{mode.signs}{A vector of length 2 (for matrices) or 3 (for tensors)
that indicates the signs of the priors (-1 for nonpositive, 1 for
nonnegative, and 0 otherwise).}

\item{tol}{Convergence tolerance.}

\item{maxiter}{Maximum number of iterations.}

\item{seed}{Since initialization is random, a default seed is set for
reproducibility.}
}
\description{
Initializes a new flash factor using \code{softImpute::softImpute}. This
  is slower for very large data matrices, but often yields better results
  when there is missing data.
}
