% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_plots.R
\name{plot.flash}
\alias{plot.flash}
\title{Plot method for flash objects}
\usage{
\method{plot}{flash}(
  x,
  include_scree = TRUE,
  include_pm = TRUE,
  order_by_pve = FALSE,
  kset = NULL,
  pm_which = c("factors", "loadings"),
  pm_subset = NULL,
  pm_groups = NULL,
  pm_colors = NULL,
  plot_type = c("scree", "bar", "heatmap", "histogram", "structure", "scatter",
    "data.frame"),
  ...
)
}
\arguments{
\item{x}{An object inheriting from class \code{flash}.}

\item{include_scree}{This parameter has been soft-deprecated; please use
\code{plot_type} instead.}

\item{include_pm}{This parameter has been soft-deprecated; please use
\code{plot_type} instead.}

\item{order_by_pve}{If \code{TRUE}, then the factor/loadings pairs will be
re-ordered according to proportion of variance explained (from
highest to lowest).}

\item{kset}{A vector of integers specifying the factor/loadings pairs to be
plotted. If \code{kset = NULL}, then all will be plotted.}

\item{pm_which}{Whether to plot loadings \eqn{L} or factors \eqn{F}.}

\item{pm_subset}{A vector of row indices \eqn{i} or column indices
\eqn{j} (depending on the argument to \code{pm_which})
specifying which values \eqn{\ell_{i \cdot}} or \eqn{f_{j \cdot}} are
to be shown. If the dataset has row or column names, then names rather
than indices may be specified. If \code{pm_subset = NULL}, then all values
will be plotted.}

\item{pm_groups}{A vector specifying the group to which each row of the data
\eqn{y_{i \cdot}} or column \eqn{y_{\cdot j}} belongs (groups may be
numeric indices or strings). A group must be provided for each plotted row
\eqn{i} or column \eqn{j}, so that the length of \code{pm_groups} is
exactly equal to the number of rows or columns in the full dataset or, if
\code{pm_subset} is specified, in the subsetted dataset.}

\item{pm_colors}{A character vector specifying a color for each unique group
specified by \code{pm_groups}, or, if \code{pm_groups = NULL}, a vector
specifying a color for each plotted row \eqn{i} or column \eqn{j}. For
effects, see parameter \code{plot_type}.}

\item{plot_type}{The type of plot to return. Options include:
\describe{
  \item{\code{"scree"}}{A scree plot showing the proportion of variance
    explained per factor/loadings pair.}
  \item{\code{"bar"}}{A simple bar plot of posterior means for loadings or
    factors (depending on argument \code{pm_which}), with one bar per
    row or column. Colors of bars are specified by argument
    \code{pm_colors}. This type of plot is most useful when rows or columns
    are small in number or ordered in a logical fashion (e.g., spatially).}
  \item{\code{"heatmap"}}{A heatmap showing posterior means for loadings or
    factors, with rows or columns grouped according to \code{pm_groups} and
    arranged within groups using the embedding provided by \code{fastTopics}
    function \code{\link[fastTopics]{structure_plot}}. Note that
    heatmaps ignore argument \code{pm_colors}.}
  \item{\code{"histogram"}}{Overlapping semi-transparent histograms of
    posterior means for loadings or factors, with one histogram per group
    specified by \code{pm_groups} (or a single histogram if \code{pm_groups}
    is \code{NULL}). Colors of histograms are specified by \code{pm_colors}.}
  \item{\code{"structure"}}{A "structure plot" (stacked bar plot) produced
    using function \code{\link[fastTopics]{structure_plot}} in package
    \code{fastTopics}. Here \code{pm_colors} specifies the colors of
    different factor/loadings pairs (as specified by \code{kset}) rather
    than different groups (as specified by \code{pm_groups}). Note that
    factors/loadings must be nonnegative for structure plots to make
    sense.}
}}

\item{...}{Additional parameters to be passed to particular
\code{flash_plot_xxx} functions. See individual functions for details.}
}
\value{
If arguments \code{include_scree} and \code{include_pm} specify that
  only one figure be produced, then \code{plot.flash()} returns a
  \code{ggplot2} object. If both figures are to be produced, then
  \code{plot.flash()} prints both plots but does not return a value.
}
\description{
Given a \code{\link{flash}} object, produces up to two figures: one showing
  the proportion of variance explained per factor/loadings pair, and one that
  plots posterior means for either factors or loadings (depending on the
  argument to parameter \code{pm_which}).
}
