% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash_plots.R
\name{plot.flash}
\alias{plot.flash}
\title{Plot method for flash objects}
\usage{
\method{plot}{flash}(
  x,
  include_scree = TRUE,
  include_pm = TRUE,
  order_by_pve = FALSE,
  kset = NULL,
  pm_which = c("factors", "loadings"),
  pm_subset = NULL,
  pm_groups = NULL,
  pm_colors = NULL,
  plot_type = c("scree", "bar", "heatmap", "histogram", "scatter", "structure"),
  ...
)
}
\arguments{
\item{x}{An object inheriting from class \code{flash}.}

\item{include_scree}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} This parameter has been deprecated; please use
\code{plot_type} instead.}

\item{include_pm}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} This parameter has been deprecated; please use
\code{plot_type} instead.}

\item{order_by_pve}{If \code{order_by_pve = TRUE}, then factor/loadings pairs
will be ordered according to proportion of variance explained, from
highest to lowest. (By default, they are plotted in the same order as
\code{kset}; or, if \code{kset} is \code{NULL}, then they are plotted in
the same order as they are found in \code{fl}.)}

\item{kset}{A vector of integers specifying the factor/loadings pairs to be
plotted. If \code{order_by_pve = FALSE}, then \code{kset} also specifies the
\emph{order} in which they are to be plotted.}

\item{pm_which}{Whether to plot loadings \eqn{L} or factors \eqn{F}.}

\item{pm_subset}{A vector of row indices \eqn{i} or column indices
\eqn{j} (depending on the argument to \code{pm_which})
specifying which values \eqn{\ell_{i \cdot}}{l_{i.}} or \eqn{f_{j \cdot}}{f_{j.}} are
to be shown. If the dataset has row or column names, then names rather
than indices may be specified. If \code{pm_subset = NULL}, then all values
will be plotted.}

\item{pm_groups}{A vector specifying the group to which each row of the data
\eqn{y_{i \cdot}}{y_{i.}} or column \eqn{y_{\cdot j}}{y_{.j}} belongs (groups may be
numeric indices or strings). A group must be provided for each plotted row
\eqn{i} or column \eqn{j}, so that the length of \code{pm_groups} is
exactly equal to the number of rows or columns in the full dataset or, if
\code{pm_subset} is specified, in the subsetted dataset.}

\item{pm_colors}{A character vector specifying a color for each unique group
specified by \code{pm_groups}, or, if \code{pm_groups = NULL}, a vector
specifying a color for each plotted row \eqn{i} or column \eqn{j}. For
effects, see parameter \code{plot_type}.}

\item{plot_type}{The type of plot to return. Options include:
\describe{
\item{\code{"scree"}}{A scree plot showing the proportion of variance
explained per factor/loadings pair. See
\code{\link{flash_plot_scree}}.}
\item{\code{"bar"}}{A bar plot of posterior means for loadings or
factors (depending on argument \code{pm_which}), with one bar per
row or column. Colors of bars are specified by argument
\code{pm_colors}. This type of plot is most useful when rows or columns
are small in number or ordered in a logical fashion (e.g., spatially).
See \code{\link{flash_plot_bar}}.}
\item{\code{"heatmap"}}{A heatmap showing posterior means for loadings or
factors, with rows or columns grouped using a 1-d embedding. Here
\code{pm_color} specifies the diverging color gradient (low-mid-high).
See \code{\link{flash_plot_heatmap}}.}
\item{\code{"histogram"}}{Overlapping semi-transparent histograms of
posterior means for loadings or factors, with one histogram per group
specified by \code{pm_groups} (or a single histogram if \code{pm_groups}
is \code{NULL}). Colors of histograms are specified by \code{pm_colors}.
See \code{\link{flash_plot_histogram}}.}
\item{\code{"scatter"}}{A scatter plot showing the relationship between
posterior means for loadings or factors and a user-supplied covariate.
If a covariate is not supplied, then data column or row means will be
used. Colors of points are specified by \code{pm_colors}. See
\code{\link{flash_plot_scatter}}.}
\item{\code{"structure"}}{A "structure plot" (stacked bar plot) produced
using function \code{\link[fastTopics]{structure_plot}} in package
\code{fastTopics}. Here \code{pm_colors} specifies the colors of
different factor/loadings pairs (as specified by \code{kset}) rather
than different groups (as specified by \code{pm_groups}). Note that
factors/loadings must be nonnegative for structure plots to make
sense. See \code{\link{flash_plot_structure}}.}
}}

\item{...}{Additional parameters to be passed to respective
\code{flash_plot_xxx} functions. See
\code{\link{flash_plot_scree}}, \code{\link{flash_plot_bar}},
\code{\link{flash_plot_heatmap}}, \code{\link{flash_plot_histogram}},
\code{\link{flash_plot_scatter}}, and \code{\link{flash_plot_structure}}
for details.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots a \code{\link{flash}} object. Several types of plot are possible:
see parameter \code{plot_type} below as well as functions
\code{\link{flash_plot_scree}}, \code{\link{flash_plot_bar}},
\code{\link{flash_plot_heatmap}}, \code{\link{flash_plot_histogram}},
\code{\link{flash_plot_scatter}}, and \code{\link{flash_plot_structure}}.
}
